BIG update.  The Python and C++ graph parser & generators are now working with a common format. The main C++ search daemon is executing searches correctly. Returned paths are being efficiently compressed to their shortest form. Refactored graph writing code into standalone class (GraphWriter.py)

+++ src/Graph.h
@@ -18,7 +18,7 @@ class Graph {
 
 	private:
 	unsigned int** adj_list;
-	unsigned short* num_outlinks; // Stores the number of outlinks per node.
+	unsigned int* num_outlinks; // Stores the number of outlinks per node.
 								// This array should be the same length as
 	unsigned int node_count;
